Cleaning testsuite from jboss.xml files - AS7 uses jboss-ejb3.xml
