+++ testsuite/integration/basic/src/test/java/org/jboss/as/test/integration/ejb/entity/cmp/cascadedelete/CustomerLocalHome.java
@@ -19,7 +19,7 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.jboss.as.test.integration.ejb.entity.cmp.cascadedelete.ejb;
+package org.jboss.as.test.integration.ejb.entity.cmp.cascadedelete;
 
 import javax.ejb.CreateException;
 import javax.ejb.EJBLocalHome;
@@ -31,9 +31,9 @@ import javax.ejb.FinderException;
  * @author <a href="alex@jboss.com">Alexey Loubyansky</a>
  * @version $Revision: 82920 $
  */
-public interface AccountLocalHome extends EJBLocalHome
+public interface CustomerLocalHome extends EJBLocalHome
 {
-   AccountLocal create(Long id, String name) throws CreateException;
+   CustomerLocal create(Long id, String name) throws CreateException;
 
-   AccountLocal findByPrimaryKey(Long id) throws FinderException;
+   CustomerLocal findByPrimaryKey(Long id) throws FinderException;
 }
