+++ testsuite/integration/basic/src/test/java/org/jboss/as/test/integration/ejb/entity/cmp/optimisticlock/CmpEntityBean.java
@@ -19,7 +19,7 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.jboss.as.test.integration.ejb.entity.cmp.optimisticlock.ejb;
+package org.jboss.as.test.integration.ejb.entity.cmp.optimisticlock;
 
 import javax.ejb.EntityBean;
 import javax.ejb.CreateException;
