+++ testsuite/integration/basic/src/test/java/org/jboss/as/test/integration/ejb/remote/entity/cmp/commerce/CascadeDeleteTestCase.java
@@ -52,7 +52,6 @@ public class CascadeDeleteTestCase {
         JavaArchive jar = ShrinkWrap.create(JavaArchive.class, MODULE_NAME + ".jar");
         jar.addPackage(CascadeDeleteTestCase.class.getPackage());
         jar.addAsManifestResource("ejb/remote/entity/cmp/commerce/ejb-jar.xml", "ejb-jar.xml");
-        jar.addAsManifestResource("ejb/remote/entity/cmp/commerce/jboss.xml", "jboss.xml");
         jar.addAsManifestResource("ejb/remote/entity/cmp/commerce/jbosscmp-jdbc.xml", "jbosscmp-jdbc.xml");
         ear.addAsModule(jar);
         return ear;
