+++ testsuite/integration/basic/src/test/java/org/jboss/as/test/integration/ejb/remote/entity/cmp/commerce/LazyResultSetLoadingTestCase.java
@@ -58,7 +58,6 @@ public class LazyResultSetLoadingTestCase {
         jar.addPackage(CommerceTestCase.class.getPackage());
         jar.addAsManifestResource("ejb/remote/entity/cmp/commerce/MANIFEST.MF", "MANIFEST.MF");
         jar.addAsManifestResource("ejb/remote/entity/cmp/commerce/ejb-jar.xml", "ejb-jar.xml");
-        jar.addAsManifestResource("ejb/remote/entity/cmp/commerce/jboss.xml", "jboss.xml");
         jar.addAsManifestResource("ejb/remote/entity/cmp/commerce/jbosscmp-jdbc.xml", "jbosscmp-jdbc.xml");
         ear.addAsModule(jar);
         return ear;
