--- testsuite/integration/basic/src/test/resources/ejb/entity/cmp/cascadedelete/jboss.xml
@@ -1,38 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE jboss PUBLIC "-//JBoss//DTD JBOSS 3.2//EN" "http://www.jboss.org/j2ee/dtd/jboss_3_2.dtd">
-
-<jboss>
-
-    <enterprise-beans>
-
-        <entity>
-            <ejb-name>Customer</ejb-name>
-            <local-jndi-name>CustomerLocal</local-jndi-name>
-            <configuration-name>custom container</configuration-name>
-        </entity>
-        <entity>
-            <ejb-name>Account</ejb-name>
-            <local-jndi-name>AccountLocal</local-jndi-name>
-            <configuration-name>custom container</configuration-name>
-        </entity>
-
-        <session>
-            <ejb-name>EJBTestRunnerEJB</ejb-name>
-            <jndi-name>ejb/EJBTestRunner</jndi-name>
-        </session>
-        <session>
-            <ejb-name>Facade</ejb-name>
-            <jndi-name>Facade</jndi-name>
-        </session>
-
-    </enterprise-beans>
-
-    <container-configurations>
-
-        <container-configuration extends="cmp2.x jdbc2 pm">
-            <container-name>custom container</container-name>
-        </container-configuration>
-
-    </container-configurations>
-
-</jboss>
