--- testsuite/integration/basic/src/test/resources/ejb/entity/cmp/commerce/jboss.xml
@@ -1,49 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<!DOCTYPE jboss PUBLIC "-//JBoss//DTD JBOSS 4.0//EN" "http://www.jboss.org/j2ee/dtd/jboss_4_0.dtd">
-
-<jboss>
-   <enterprise-beans>
-      <entity>
-         <ejb-name>UserEJB</ejb-name>
-         <jndi-name>commerce/User</jndi-name>
-         <local-jndi-name>commerce/UserLocalHome</local-jndi-name>
-      </entity>
-      <entity>
-         <ejb-name>CustomerEJB</ejb-name>
-         <jndi-name>commerce/Customer</jndi-name>
-      </entity>
-      <entity>
-         <ejb-name>OrderEJB</ejb-name>
-         <local-jndi-name>commerce/Order</local-jndi-name>
-      </entity>
-      <entity>
-         <ejb-name>AddressEJB</ejb-name>
-         <local-jndi-name>commerce/Address</local-jndi-name>
-      </entity>
-      <entity>
-         <ejb-name>LineItemEJB</ejb-name>
-         <local-jndi-name>commerce/LineItem</local-jndi-name>
-      </entity>
-      <entity>
-         <ejb-name>ProductEJB</ejb-name>
-         <local-jndi-name>commerce/Product</local-jndi-name>
-      </entity>
-      <entity>
-         <ejb-name>ProductCategoryEJB</ejb-name>
-         <local-jndi-name>commerce/ProductCategory</local-jndi-name>
-      </entity>
-      <entity>
-         <ejb-name>ProductCategoryTypeEJB</ejb-name>
-         <local-jndi-name>commerce/ProductCategoryType</local-jndi-name>
-      </entity>
-      <entity>
-         <ejb-name>ProductCategoryBatchDeleteEJB</ejb-name>
-         <local-jndi-name>commerce/ProductCategoryBatchDelete</local-jndi-name>
-      </entity>
-      <entity>
-         <ejb-name>ProductCategoryTypeBatchDeleteEJB</ejb-name>
-         <local-jndi-name>commerce/ProductCategoryTypeBatchDelete</local-jndi-name>
-      </entity>
-   </enterprise-beans>
-</jboss>
