--- testsuite/integration/basic/src/test/resources/ejb/entity/cmp/optimisticlock/jboss.xml
@@ -1,72 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE jboss PUBLIC "-//JBoss//DTD JBOSS 3.0//EN" "http://www.jboss.org/j2ee/dtd/jboss_3_0.dtd">
-
-<jboss>
-   <enterprise-beans>
-      <entity>
-         <ejb-name>EntityGroupLocking</ejb-name>
-         <local-jndi-name>local/EntityGroupLocking</local-jndi-name>
-         <configuration-name>Optimistic Container</configuration-name>
-      </entity>
-      <entity>
-         <ejb-name>EntityModifiedLocking</ejb-name>
-         <local-jndi-name>local/EntityModifiedLocking</local-jndi-name>
-         <configuration-name>Optimistic Container</configuration-name>
-      </entity>
-      <entity>
-         <ejb-name>EntityReadLocking</ejb-name>
-         <local-jndi-name>local/EntityReadLocking</local-jndi-name>
-         <configuration-name>Optimistic Container</configuration-name>
-      </entity>
-      <entity>
-         <ejb-name>EntityVersionLocking</ejb-name>
-         <local-jndi-name>local/EntityVersionLocking</local-jndi-name>
-         <configuration-name>Optimistic Container</configuration-name>
-      </entity>
-      <entity>
-         <ejb-name>EntityExplicitVersionLocking</ejb-name>
-         <local-jndi-name>local/EntityExplicitVersionLocking</local-jndi-name>
-         <configuration-name>Optimistic Container</configuration-name>
-      </entity>
-      <entity>
-         <ejb-name>EntityTimestampLocking</ejb-name>
-         <local-jndi-name>local/EntityTimestampLocking</local-jndi-name>
-         <configuration-name>Optimistic Container</configuration-name>
-      </entity>
-      <entity>
-         <ejb-name>EntityKeyGeneratorLocking</ejb-name>
-         <local-jndi-name>local/EntityKeyGeneratorLocking</local-jndi-name>
-         <configuration-name>Optimistic Container</configuration-name>
-      </entity>
-
-		<entity>
-			<ejb-name>EntityA</ejb-name>
-			<configuration-name>Optimistic Container</configuration-name>
-		</entity>
-		<entity>
-			<ejb-name>EntityB</ejb-name>
-			<configuration-name>Optimistic Container</configuration-name>
-		</entity>
-		<session>
-			<ejb-name>TestSession</ejb-name>
-			<jndi-name>ejb/TestSession</jndi-name>
-		</session>
-
-      <session>
-         <ejb-name>EJBTestRunnerEJB</ejb-name>
-         <jndi-name>ejb/EJBTestRunner</jndi-name>
-      </session>
-      <session>
-         <ejb-name>Facade</ejb-name>
-         <jndi-name>FacadeBean</jndi-name>
-      </session>
-   </enterprise-beans>
-
-   <container-configurations>
-      <container-configuration extends="Instance Per Transaction CMP 2.x EntityBean">
-         <container-name>Optimistic Container</container-name>
-         <!-- <locking-policy>org.jboss.ejb.plugins.lock.JDBCOptimisticLock</locking-policy> -->
-         <locking-policy>org.jboss.ejb.plugins.lock.NoLock</locking-policy>
-      </container-configuration>
-   </container-configurations>
-</jboss>
