--- testsuite/integration/basic/src/test/resources/ejb/entity/cmp/relationship/jboss.xml
@@ -1,120 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!DOCTYPE jboss PUBLIC "-//JBoss//DTD JBOSS 4.0//EN" "http://www.jboss.org/j2ee/dtd/jboss_4_0.dtd">
-
-<jboss>
-   <enterprise-beans>
-      <session>
-         <ejb-name>EJBTestRunnerEJB</ejb-name>
-         <jndi-name>ejb/EJBTestRunner</jndi-name>
-      </session>
-
-      <entity>
-         <ejb-name>A_OneToOne_Bi_Table_EJB</ejb-name>
-         <local-jndi-name>relation/oneToOne/bidirectional/table/A</local-jndi-name>
-      </entity>
-      <entity>
-         <ejb-name>B_OneToOne_Bi_Table_EJB</ejb-name>
-         <local-jndi-name>relation/oneToOne/bidirectional/table/B</local-jndi-name>
-      </entity>
-
-      <entity>
-         <ejb-name>A_OneToOne_Bi_FK_EJB</ejb-name>
-         <local-jndi-name>relation/oneToOne/bidirectional/fk/A</local-jndi-name>
-      </entity>
-      <entity>
-         <ejb-name>B_OneToOne_Bi_FK_EJB</ejb-name>
-         <local-jndi-name>relation/oneToOne/bidirectional/fk/B</local-jndi-name>
-      </entity>
-
-      <entity>
-         <ejb-name>A_OneToOne_Uni_Table_EJB</ejb-name>
-         <local-jndi-name>relation/oneToOne/unidirectional/table/A</local-jndi-name>
-      </entity>
-      <entity>
-         <ejb-name>B_OneToOne_Uni_Table_EJB</ejb-name>
-         <local-jndi-name>relation/oneToOne/unidirectional/table/B</local-jndi-name>
-      </entity>
-
-      <entity>
-         <ejb-name>A_OneToOne_Uni_FK_EJB</ejb-name>
-         <local-jndi-name>relation/oneToOne/unidirectional/fk/A</local-jndi-name>
-      </entity>
-      <entity>
-         <ejb-name>B_OneToOne_Uni_FK_EJB</ejb-name>
-         <local-jndi-name>relation/oneToOne/unidirectional/fk/B</local-jndi-name>
-      </entity>
-
-      <entity>
-         <ejb-name>A_OneToMany_Bi_Table_EJB</ejb-name>
-         <local-jndi-name>relation/oneToMany/bidirectional/table/A</local-jndi-name>
-      </entity>
-      <entity>
-         <ejb-name>B_OneToMany_Bi_Table_EJB</ejb-name>
-         <local-jndi-name>relation/oneToMany/bidirectional/table/B</local-jndi-name>
-      </entity>
-
-      <entity>
-         <ejb-name>A_OneToMany_Bi_FK_EJB</ejb-name>
-         <local-jndi-name>relation/oneToMany/bidirectional/fk/A</local-jndi-name>
-      </entity>
-      <entity>
-         <ejb-name>B_OneToMany_Bi_FK_EJB</ejb-name>
-         <local-jndi-name>relation/oneToMany/bidirectional/fk/B</local-jndi-name>
-      </entity>
-
-      <entity>
-         <ejb-name>A_OneToMany_Uni_Table_EJB</ejb-name>
-         <local-jndi-name>relation/oneToMany/unidirectional/table/A</local-jndi-name>
-      </entity>
-      <entity>
-         <ejb-name>B_OneToMany_Uni_Table_EJB</ejb-name>
-         <local-jndi-name>relation/oneToMany/unidirectional/table/B</local-jndi-name>
-      </entity>
-
-      <entity>
-         <ejb-name>A_OneToMany_Uni_FK_EJB</ejb-name>
-         <local-jndi-name>relation/oneToMany/unidirectional/fk/A</local-jndi-name>
-      </entity>
-      <entity>
-         <ejb-name>B_OneToMany_Uni_FK_EJB</ejb-name>
-         <local-jndi-name>relation/oneToMany/unidirectional/fk/B</local-jndi-name>
-      </entity>
-
-      <entity>
-         <ejb-name>A_ManyToOne_Uni_Table_EJB</ejb-name>
-         <local-jndi-name>relation/manyToOne/unidirectional/table/A</local-jndi-name>
-      </entity>
-      <entity>
-         <ejb-name>B_ManyToOne_Uni_Table_EJB</ejb-name>
-         <local-jndi-name>relation/manyToOne/unidirectional/table/B</local-jndi-name>
-      </entity>
-
-      <entity>
-         <ejb-name>A_ManyToOne_Uni_FK_EJB</ejb-name>
-         <local-jndi-name>relation/manyToOne/unidirectional/fk/A</local-jndi-name>
-      </entity>
-      <entity>
-         <ejb-name>B_ManyToOne_Uni_FK_EJB</ejb-name>
-         <local-jndi-name>relation/manyToOne/unidirectional/fk/B</local-jndi-name>
-      </entity>
-
-      <entity>
-         <ejb-name>A_ManyToMany_Bi_EJB</ejb-name>
-         <local-jndi-name>relation/manyToMany/bidirectional/A</local-jndi-name>
-      </entity>
-      <entity>
-         <ejb-name>B_ManyToMany_Bi_EJB</ejb-name>
-         <local-jndi-name>relation/manyToMany/bidirectional/B</local-jndi-name>
-      </entity>
-
-      <entity>
-         <ejb-name>A_ManyToMany_Uni_EJB</ejb-name>
-         <local-jndi-name>relation/manyToMany/unidirectional/A</local-jndi-name>
-      </entity>
-      <entity>
-         <ejb-name>B_ManyToMany_Uni_EJB</ejb-name>
-         <local-jndi-name>relation/manyToMany/unidirectional/B</local-jndi-name>
-      </entity>
-
-   </enterprise-beans>
-</jboss>
