--- testsuite/integration/basic/src/test/resources/ejb/entity/cmp/simple/jboss.xml
@@ -1,33 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!DOCTYPE jboss PUBLIC
-      "-//JBoss//DTD JBOSS 3.2//EN"
-      "http://www.jboss.org/j2ee/dtd/jboss_3_2.dtd">
-
-<jboss>
-   <enterprise-beans>
-      <session>
-         <ejb-name>EJBTestRunnerEJB</ejb-name>
-         <jndi-name>ejb/EJBTestRunner</jndi-name>
-      </session>
-
-      <entity>
-         <ejb-name>SimpleEJB</ejb-name>
-         <local-jndi-name>cmp2/simple/Simple</local-jndi-name>
-      </entity>
-
-      <entity>
-         <ejb-name>SimpleEJBWithOptionA</ejb-name>
-         <local-jndi-name>cmp2/simple/SimpleA</local-jndi-name>
-         <read-only>false</read-only>
-         <configuration-name>CMP 2.x and Cache</configuration-name>
-      </entity>
-   </enterprise-beans>
-
-   <container-configurations>
-      <container-configuration extends="Standard CMP 2.x EntityBean">
-         <container-name>CMP 2.x and Cache</container-name>
-         <commit-option>A</commit-option>
-      </container-configuration>
-   </container-configurations>
-
-</jboss>
